pdf generation for beleg included

+++ grails-app/domain/hierlmeier/Positionstyp.groovy
@@ -2,7 +2,7 @@ package hierlmeier
 
 class Positionstyp implements Serializable { //muss Seriazable implementieren für Flows in Grails
     
-    Integer preis
+    BigDecimal preis
     String bezeichnung
 
     static constraints = {
