+++ grails-app/domain/hierlmeier/Zahlungsteil.groovy
@@ -2,12 +2,15 @@ package hierlmeier
 
 class Zahlungsteil {
     
-    Integer betrag      // == teilbetrag einer Zahlung
+    BigDecimal betrag      // == teilbetrag einer Zahlung
     
     Zahlung zahlung
     Beleg beleg
 
     static constraints = {
+        //betrag(scale:2) /*@todo checken wo ma wie die nachkommastellen checkt 
+        //weil hier in den constraints, bedeutet das nur das abgeschnitte wird,
+        //und kein fehler geschmissen
     }
     
     def String toString() {
